refactor: Admin 관련 클래스 및 경로 리팩토링
- AdminReservation 관련 클래스와 요청 경로를 admin 패키지로 이동
- AdminReservationController의 endpoint를 "/admin"에서 "/admin/reservations"로 수정
- 관련 import문과 테스트 코드 경로를 변경하여 패키지 변경 사항 반영
