+++ src/test/java/roomescape/reservation/application/ReservationServiceTest.java
@@ -14,12 +14,12 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ActiveProfiles;
-import roomescape.reservation.time.application.ReservationTimeRepository;
-import roomescape.theme.application.ThemeRepository;
-import roomescape.reservation.time.domain.ReservationTime;
-import roomescape.theme.domain.Theme;
-import roomescape.reservation.presentation.dto.AdminReservationRequest;
+import roomescape.admin.AdminReservationRequest;
 import roomescape.reservation.presentation.dto.ReservationResponse;
+import roomescape.reservation.time.application.ReservationTimeRepository;
+import roomescape.reservation.time.domain.ReservationTime;
+import roomescape.theme.application.ThemeRepository;
+import roomescape.theme.domain.Theme;
 
 @ActiveProfiles("test")
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
