+++ src/test/java/roomescape/reservation/presentation/AdminReservationControllerTest.java
@@ -10,8 +10,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ActiveProfiles;
+import roomescape.admin.AdminReservationRequest;
 import roomescape.member.presentation.fixture.MemberFixture;
-import roomescape.reservation.presentation.dto.AdminReservationRequest;
 import roomescape.reservation.presentation.fixture.ReservationFixture;
 
 @ActiveProfiles("test")
